% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_artoo_checks}
\alias{is_artoo_checks}
\title{Test for a artoo_checks control}
\usage{
is_artoo_checks(x)
}
\arguments{
\item{x}{\emph{Object to test.} \verb{<any>}.}
}
\value{
\emph{A \verb{<logical(1)>}}: \code{TRUE} when \code{x} is a \code{artoo_checks}.
}
\description{
Report whether an object is a \code{artoo_checks} control built by
\code{\link[=artoo_checks]{artoo_checks()}}. Use it to guard a \code{checks} argument before threading it
into \code{\link[=check_spec]{check_spec()}} or \code{\link[=apply_spec]{apply_spec()}}.
}
\examples{
# ---- Example 1: confirm a control before reusing it ----
#
# is_artoo_checks() distinguishes a real control from a bare list of flags.
is_artoo_checks(artoo_checks())
is_artoo_checks(list(missing_variable = TRUE))

}
\seealso{
\code{\link[=artoo_checks]{artoo_checks()}} to build one.
}
