% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_construct.R
\name{is_vport_spec}
\alias{is_vport_spec}
\title{Is \code{x} a vport_spec?}
\usage{
is_vport_spec(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{vport_spec}, otherwise \code{FALSE}.
}
\description{
Is \code{x} a vport_spec?
}
\examples{
is_vport_spec(vport_spec(
  data.frame(dataset = "DM"),
  data.frame(dataset = "DM", variable = "AGE", data_type = "integer")
))
is_vport_spec(mtcars)
}
\seealso{
\code{\link[=vport_spec]{vport_spec()}}.
}
